Refactor directories, add components and service.
